requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohttp==3.11.13
bitsandbytes # ==0.45.3
deepspeed==0.16.4
diffusers==0.35.2
einops==0.8.1
huggingface-hub==0.36.0
matplotlib==3.10.0
ml_collections==1.1.0
numpy==1.26.4
openai==2.6.1
peft==0.17.1
pillow==11.3.0
pydantic
pydantic_core
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.1
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
uvicorn==0.34.0
wandb==0.18.7
xformers
mmdet==2.28.2
open_clip_torch
clip-benchmark
mmcv-full==1.7.2
# for ocr
paddlepaddle_gpu==3.1.1
paddleocr==3.2.0
paddlex==3.2.1
opencv_contrib_python==4.10.0.84
python-Levenshtein
icecream
fairscale
einops-exts
open_clip_torch
ftfy
timm
shortuuid
qwen-vl-utils